New and improved Tasks semantics
This mostly-rewrites the rest of Tasks, so that we can offer improved one-shot task semantics. By default, tasks now auto-clear their pending status. You may retrigger them, and the same or other tasks can reliably cancel a triggered task. The AudioPull task is now one-shot, and we trigger it only as a result of handing off some audio data to the output device.

As a result, all of our normal system behavior (when we're "idle") happens with zero pending tasks. If a task is pending, we expect to perform some quantity of work and then the task will no longer be pending.

So, we can *finally* implement the semantics that Tasks::idle() does not call WFI unless no other work was performed. This should help performance, as we won't be waiting when we don't need to.