Adjust radio address calculations to avoid illegal channels
Liam pointed out that channels above 83 are supported by the silicon, but not legal to transmit on without a license in the US. So, I've updated the algorithms to use this alternate upper limit. This included picking a new polynomial for the pairing channel algorithm.

While I was in there, I fixed a flaw in the HWID-based address generator which was causing channel numbers to have no additional entropy, meaning that the channel was always predictable if you knew the last byte of the address.