Protocol change, accelerometer signs match hardware
Since we're breaking protocol in a big way anyway, taking this opportunity to make the wire format for accelerometer readings match the hardware again. This saves a little code space.

Also documented the SDK's standard coordinate system.

Also: The way we were inverting the readings (cpl) is not quite correct. That's a one's complement, not a two's complement. So this may fix some off-by-one with negative acceleration.