Reuse system cube IDs even when userspace hasn't seen disconnection yet
This patch changes the way we handle cube disconnect events a bit. We actually do *not* want to avoid reallocating IDs that userspace may still be using- it's okay to do so, and is required to ensure that our cube IDs stay in range. This also makes our disconnect event handling more deliberate, so we can effectively dispatch disconnect events even when a cube has reconnected in the same slot.