Siftulator audio: Prevent output buffer starvation
If the audio device's buffer was larger than our mixing buffer, we could fail to keep it full due to our limited ability to pull samples from the smaller buffer to the larger one only during audio callbacks. This change introduces a hook that the mixer uses to let us pull audio directly, in task context, in a way that lets us immediately re-trigger the task rather than waiting for an audio device callback.

This change is an artifact of one big difference between siftulator and hardware: On our hardware, the output device operates on units (individual samples) that are much smaller than the mix buffer, whereas in siftulator the units we operate on (operating system sample buffers) may be much larger.