New base lid artwork
Whew. I was so not happy with the aliasing from the high poly model. So, let's do it right. Ambient occlusion map that took like 45 minutes to render on my lappy. Fuck yeah. Also a normal map. And we can do a nice shiny LED in the fragment shader too.

Includes multiple layers in the .blend for cube, base, base lid, and simplified base lid. The model_mc_face resource has been replaced with the simplified lid.