Meh... Cleaned up cube body normals and lightmap
Still not happy with this. These meshes were in kinda bad shape... the top frame of the cube has some crunchy bits that I mostly gave up on, and ended up blurring a bit in photoshop. It still looks crappy. And so far I haven't found a good way to make a low-res mesh for this cube while preserving UVs.