Fix pairing delays caused by HopConfirm failure
Pairing was taking way longer than it was supposed to, because we'd prematurely fail in HopConfirm if we got a short but non-empty ACK response. This can happen easily if the cube already buffered this short ACK between handling the Radio Hop and getting the Full ACK Request token. We should resend the full ACK request in this case, not drop the cube.