Support for out-of-band VRAM writes
This is one of the prereqs for the new pause menu. I tried using a VideoBuffer to draw to each connected cube in sequence. With some tricks to prime the paintcontroller I managed to make it suck only moderately, but it still kind of sucks. This patch introduces a different approach, where we can assemble special-purpose VRAM updates directly in the ISR as long as they fit into a single packet.