Cube firmware: Pulse stretching for touch events
This change does not break protocol. We detect touch events
slightly differently: In TF0, if a touch is detected we
force an ACK packet transmission with the touch bit set,
but we never clear the touch bit until after an ACK transmit
has finished. This ensures that any touch (a) is not going to
be missed entirely, and (b) we spam out many ACKs while the touch
is being detected, making touch more resilient to radio glitches.

(This affects the touch *down* event only. Touch *up* is no more
robust after this change, though we could also detect lost ACKs
there via a watchdog timer. I.e. if we've received more than a
couple empty ACKs or ACKs without neighbor bits, we know the touch
has ended.)