New Events::gameMenu in SDK
This adds a way to set up an event handler for a custom menu item in
the standard "pause" menu. This is totally optional, and if you don't
set it up the pause menu will only have continue and quit options. If
you do set it up, this event will fire when the custom item is chosen,
and you get to pick a label for the item.