Initial checkin for Shutdown manager
This adds a new module, ShutdownManager, and a bunch of related infrastructure. This thing handles orchestrating the process of both user-visible and hardware shutdown, and it gives us a place to run tasks that happen while the base is "off", like USB interaction, shutting down cubes asynchronously, and doing LFS garbage collection and flash pre-erase.