Fix overzealous audio buffer pre-fill on Siftulator
Resolves https://sifteo.atlassian.net/browse/THCR-54

This caused a potentially very large buffering delay at the very beginning of
an audio stream- such as starting a sound effect while music is not already
playing. We have an adaptive threshold (which could grow to be fairly large)
to control how much audio we try to keep in the buffer. But we were using that
same threshold to determine how much audio needs to already be buffered before
we start playing a new stream, to avoid skipping at the very beginning of
the stream. Because of the adaptive threshold, this could now be much too
large for that purpose. The patch simply uses a separate fixed threshold for
pre-fill, instead of using the dynamic threshold.