Add a basic cubeRefresh handler to text demo
This doesn't redraw especially gracefully at the moment, but it's a good example for what I hope minimal but complete refresh handling in userspace looks like. I think the system needs to be more careful about the order in which restoring from the home menu happens- like we could ensure that the refresh handler gets run before the first Paint.