Detect exploitable ROP gadgets in Cube firmware
DJ nerdsniped me into thinking about this attack vector. If we assume there's already an exploit that allows an attacker to control the contents of the stack (or write to IRAM in general), this is the last line of defense preventing them from bricking a cube by writing over OTP memory. So far we're mostly good, but there are a couple of exploitable code sequences in the tilerom.