AssetGroup ABI change: LUT contents are undefined initially
This makes it much more efficient for AssetLoader to stitch together several groups. It means we only need to send one Reset per load operation, and sending several AssetGroups in rapid succession can be done with little performance penalty.

In practice, this will only add a couple bytes to the size of each assetgroup, and even then it'll only do so if the color black appears in one of the first few tiles in the group.