Fix incomplete RLE4 initialization in reference LoadstreamDecoder
This fixes one of the asset-related bugs that Andrew found during game jam. This was a bit obscure, requiring several conditions for it to be noticeable:

 1. Asset Loader Bypass is in use
 2. The AssetGroup needs to begin with an RLE4 tile, not a 16-bit tile
 3. The uninitialized value of rle2 needed to be exactly 0xFF. We were only unlucky enough to observe this so far on 32-bit linux, and only on optimized builds.

In the RLE4 setup, somehow our initialization for the rle2 state got lost. This wasn't usually a big deal, since the initial value only acts as a sentinel to ensure that we don't get a false RLE match. But in this unlucky circumstance, we did get just that, and it caused at least one spurious pixel to be emitted at the very beginning of the load, throwing everything else out of alignment.