New two-stage asset FIFO reset
This allows us to be totally certain that the device's FIFO is empty when we initialize our flow control, even if we're doing a reset while the cube may still be busy programming flash. This fixes our retry path, where if a load fails for whatever reason we can automatically and immediately restart it.