Cut default software retry count in half
This was so epically long that, when one cube disconnected (timed out), it would cause a long enough stall in all other cubes that those cubes would time out while trying to load assets. That then exposed yet another bug in the recovery process. Now that the fix for the recovery process is in, this patch fixes the source of these secondary timeouts.