Siftulator support for wake on shake and RF
This is a big refactor of the powerdown code in siftulator, so that we can support modes other than deep sleep. Now the CPU core proper and the Cube::Hardware both keep running while asleep, but the CPU disables instruction execution if it's powered down, and only runs timers if timers are on. The watchdog can keep running (and generating resets) even during sleep, and there's an external wakeup source in the CPU now which Cube::Hardware uses to implement wake-on-pin with the new LIS3DH interrupt output.