Shiny new SvmValidator implementation!
After many bagels and showers and long contemplative walks, and also some coding, here now appears an SvmValidator which should actually work!

Also it's fast! It can almost always run in a single pass, and the resulting compiled code is very compact.

So far, I know it compiles and it runs the launcher successfully. And the output I've spot-checked looks good. But my next job is writing some significantly more serious unit tests for this. The validator absolutely needs to be right, or we risk breaking our ABI later in order to fix security bugs.