Do not branch() on code mapping failure
If we failed to map a code address, we were still branching
there after the fault handler returns. We should only do one or
the other- the fault handler needs to not return at all, or it needs
to clean up the CPU state appropriately by branching/exec/whatevs.

This was breaking my unit test because the handler's branch() was getting clobbered by the totally bogus branch.