Mixer: Support for non-unity mixer gain
This change sets us up to support a "maximum" volume which corresponds to a gain other than 0 dB. Currently the gain is set to 6 dB, which seems like a good level to start with. Max gains greater than 0 dB make a limiter very important, and currently the soft limiter I plan to use is just a stub. If you increase volume in Siftulator, you'll start to hear some clipping on most games. The soft limiter will help with this once it's in.