Fix for large unsigned 32-bit ints in SVMDecorations::testAndStripNumberedPrefix()
The new test in testAndStripNumberedPrefix() was giving a false error for
offsets which are large unsigned 32-bit numbers that can't be represented in
a signed 32-bit integer. To get around this, we'll do the conversion as a 64-bit
signed int, then do our own (looser) test when converting back to 32 bit.

Unit tests work again~