Remove incorrect sysHandlerPrioritize() call
This is an actual bug I found while fixing warnings after upgrading to gcc 4.7.3. There is no priority register for the HardFault handler, so this actually ends up calculating a totally bogus address. The new compiler is smart enough to know this at compile-time, so it complains about an out-of-bound array access.

If I'm calculating it correctly, the buggy code here was actually writing one bit to an arbitrary location near the top of user RAM. This is harmless so early-on in init, but if this line were called while a game is running, we'd likely be smashing that game's stack.