Drain the nRF8001 event queue before reset
I thought this was the root cause behind a bug in which connections never complete. I've found two slightly different initialization sequences, one which triggers the bug and one which doesn't. This was one of the differences that I thought was responsible, but it doesn't seem to help.

Checking this in anyway though, since it seems like a good idea. At powerup, the nRF8001 usually has a mode change message waiting for us.